// cpo-server/src/types/Setting.ts
export enum TechnicalSettings {
  CRYPTO = 'crypto',
  USER = 'user',
  TASK = 'task',
}
// ...
export interface Setting extends SettingAuthorizationActions, CreatedUpdatedProps {
  id?: string;
  identifier: TenantComponents | TechnicalSettings;
  sensitiveData?: string[];
  backupSensitiveData?: Record<string, unknown>;
  category?: 'business' | 'technical';
}
// ...
export enum PricingSettingsType {
  SIMPLE = 'simple',
}

export interface PricingSettings extends Setting {
  identifier: TenantComponents.PRICING;
  type: PricingSettingsType;
  simple?: SimplePricingSetting;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface PricingSetting {
}

export enum PricingContentType {
  SIMPLE = 'simple',
}

export interface SimplePricingSetting extends PricingSetting {
  price: number;
  currency: string;
  last_updated?: Date;
}